backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate
Flask-Cors
Flask-JWT-Extended
Flask-SocketIO
python-dotenv
mysql-connector-python
gunicorn
